+++ android/src/main/kotlin/com/hiennv/flutter_callkit_incoming/SharedPreferencesUtils.kt
@@ -1,11 +1,11 @@
-package com.khailv.flutter_callkit_incoming
+package com.hiennv.flutter_callkit_incoming
 
 import android.content.Context
 import android.content.SharedPreferences
 import com.google.gson.reflect.TypeToken
 
 
-private const val CALLKIT_PREFERENCES_FILE_NAME = "meeyteam_callkit_incoming"
+private const val CALLKIT_PREFERENCES_FILE_NAME = "flutter_callkit_incoming"
 private var prefs: SharedPreferences? = null
 private var editor: SharedPreferences.Editor? = null
 
@@ -15,11 +15,16 @@ private fun initInstance(context: Context) {
 }
 
 
-fun addCall(context: Context?, data: Data) {
+fun addCall(context: Context?, data: Data, isAccepted: Boolean = false) {
     val json = getString(context, "ACTIVE_CALLS", "[]")
     val arrayData: ArrayList<Data> = Utils.getGsonInstance()
         .fromJson(json, object : TypeToken<ArrayList<Data>>() {}.type)
-    arrayData.add(data)
+    val currentData = arrayData.find { it == data }
+    if(currentData != null) {
+        currentData.isAccepted = isAccepted
+    }else {
+        arrayData.add(data)
+    }
     putString(context, "ACTIVE_CALLS", Utils.getGsonInstance().toJson(arrayData))
 }